requirements.txt
aiofiles==24.1.0
aiosmtplib==3.0.2
asgiref==3.8.1
asyncio==3.4.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
greenlet==3.1.1
greenletio==0.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
motor==3.6.0
priority==2.0.0
pymongo==4.9.2
python-dotenv==1.0.1
Quart==0.19.9
quart-cors==0.7.0
requests==2.32.3
urllib3==2.2.3
uvicorn==0.32.1
Werkzeug==3.1.3
wsproto==1.2.0
